Updates for easier publishing

+++ gendoc.sh
@@ -65,13 +65,15 @@ done
 
 echo "- done processing opts"
 
-echo "updating dependencies by rm/checkout"
-rm -r rm -rf ${script_root}/docfx_project/ziti-*
-git clone https://github.com/openziti/ziti --branch release-next --single-branch docfx_project/ziti-cmd
-git clone https://github.com/openziti/ziti-sdk-csharp --branch main --single-branch docfx_project/ziti-sdk-csharp
-git clone https://github.com/openziti/ziti-sdk-c --branch main --single-branch docfx_project/ziti-sdk-c
-git clone https://github.com/netfoundry/ziti-android-app --branch main --single-branch docfx_project/ziti-android-app
-git clone https://github.com/openziti/ziti-sdk-swift --branch main --single-branch docfx_project/ziti-sdk-swift
+if [[ ! "${SKIP_GIT}" == "yes" ]]; then
+  echo "updating dependencies by rm/checkout"
+  rm -r rm -rf ${script_root}/docfx_project/ziti-*
+  git clone https://github.com/openziti/ziti --branch release-next --single-branch docfx_project/ziti-cmd
+  git clone https://github.com/openziti/ziti-sdk-csharp --branch main --single-branch docfx_project/ziti-sdk-csharp
+  git clone https://github.com/openziti/ziti-sdk-c --branch main --single-branch docfx_project/ziti-sdk-c
+  git clone https://github.com/netfoundry/ziti-android-app --branch main --single-branch docfx_project/ziti-android-app
+  git clone https://github.com/openziti/ziti-sdk-swift --branch main --single-branch docfx_project/ziti-sdk-swift
+fi
 
 DOC_ROOT=docs-local
